improve match collection

+++ data_collect/match.py
@@ -22,18 +22,24 @@ def collect_match():
         for match_history in \
             mypymongo.db.player_seed_match_history.find({
                                                 'accountId': account_id,
+                                                "season": 11,
+                                                "queue": 420,    # only ranked match solo queue
                                                 '$or': [
                                                     {'match_history_in_match': None},
                                                     {'match_history_in_match': False}
                                                 ]}):
             match = Match(id=match_history['gameId'], region="NA")
-            match.timeline    # call time line to fetch time line
-            mypymongo.insert_match(match.to_json())
+
+            if not mypymongo.exist_match_id_in_match(match_history['gameId']):
+                a = match.duration  # call duration to fetch real match data
+                # match.timeline    # call time line to fetch time line
+                mypymongo.insert_match(match.to_json())
 
             match_history['match_history_in_match'] = True
             mypymongo.db.player_seed_match_history.find_one_and_replace(
                 {'accountId': account_id, 'gameId': match_history['gameId']}, match_history)
-            print("finish crawling match", match_history['gameId'])
+            print("finish crawling match", match_history['gameId'], 'for', account_id,
+                  '. ', match.season, match.queue, match.creation, match.version)
 
         player['player_in_match'] = True
         mypymongo.db.player_seed.find_one_and_replace({'accountId': account_id}, player)
@@ -42,7 +48,7 @@ def collect_match():
 
 
 if __name__ == '__main__':
-    myconfig = MyConfig()
+    myconfig = MyConfig('config_chen.txt')
     config = cass.get_default_config()
     config['logging']['print_riot_api_key'] = True
     config['pipeline']['RiotAPI']['api_key'] = myconfig.riot_api_key
